demo/requirements.txt
Appium-Python-Client>=0.47
pytest
